实时互动视频游戏/requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.0.2
langgraph>=0.0.20
openai>=1.0.0
python-dotenv>=1.0.0
opencv-python>=4.8.0
requests>=2.31.0
google-genai>=0.2.0
Pillow>=10.0.0

